Renames the modules
