+++ lib/settings.py
